Split MQTT into up/down/events
