+++ mqtt/client.go
@@ -4,7 +4,6 @@
 package mqtt
 
 import (
-	"encoding/json"
 	"fmt"
 	"sync"
 	"time"
@@ -44,6 +43,14 @@ type Client interface {
 	UnsubscribeDeviceDownlink(appID string, devID string) Token
 	UnsubscribeAppDownlink(appID string) Token
 	UnsubscribeDownlink() Token
+
+	// Event pub/sub
+	PublishAppEvent(appID string, eventType string, payload interface{}) Token
+	PublishDeviceEvent(appID string, devID string, eventType string, payload interface{}) Token
+	SubscribeAppEvents(appID string, eventType string, handler AppEventHandler) Token
+	SubscribeDeviceEvents(appID string, devID string, eventType string, handler DeviceEventHandler) Token
+	UnsubscribeAppEvents(appID string, eventType string) Token
+	UnsubscribeDeviceEvents(appID string, devID string, eventType string) Token
 
 	// Activation pub/sub
 	PublishActivation(payload Activation) Token
@@ -129,15 +136,6 @@ func (t *token) Error() error {
 	defer t.RUnlock()
 	return t.err
 }
-
-// UplinkHandler is called for uplink messages
-type UplinkHandler func(client Client, appID string, devID string, req UplinkMessage)
-
-// DownlinkHandler is called for downlink messages
-type DownlinkHandler func(client Client, appID string, devID string, req DownlinkMessage)
-
-// ActivationHandler is called for activations
-type ActivationHandler func(client Client, appID string, devID string, req Activation)
 
 // DefaultClient is the default MQTT client for The Things Network
 type DefaultClient struct {
@@ -233,6 +231,10 @@ func (c *DefaultClient) Connect() error {
 	return nil
 }
 
+func (c *DefaultClient) publish(topic string, msg []byte) Token {
+	return c.mqtt.Publish(topic, QoS, false, msg)
+}
+
 func (c *DefaultClient) subscribe(topic string, handler MQTT.MessageHandler) Token {
 	c.subscriptions[topic] = handler
 	return c.mqtt.Subscribe(topic, QoS, handler)
@@ -255,235 +257,4 @@ func (c *DefaultClient) Disconnect() {
 // IsConnected returns true if there is a connection to the MQTT broker
 func (c *DefaultClient) IsConnected() bool {
 	return c.mqtt.IsConnected()
-}
-
-// PublishUplink publishes an uplink message to the MQTT broker
-func (c *DefaultClient) PublishUplink(dataUp UplinkMessage) Token {
-	topic := DeviceTopic{dataUp.AppID, dataUp.DevID, Uplink, ""}
-	dataUp.AppID = ""
-	dataUp.DevID = ""
-	msg, err := json.Marshal(dataUp)
-	if err != nil {
-		return &simpleToken{fmt.Errorf("Unable to marshal the message payload")}
-	}
-	return c.mqtt.Publish(topic.String(), QoS, false, msg)
-}
-
-// PublishUplinkFields publishes uplink fields to MQTT
-func (c *DefaultClient) PublishUplinkFields(appID string, devID string, fields map[string]interface{}) Token {
-	flattenedFields := make(map[string]interface{})
-	flatten("", "/", fields, flattenedFields)
-	tokens := make([]Token, 0, len(flattenedFields))
-	for field, value := range flattenedFields {
-		topic := DeviceTopic{appID, devID, Uplink, field}
-		pld, _ := json.Marshal(value)
-		token := c.mqtt.Publish(topic.String(), QoS, false, pld)
-		tokens = append(tokens, token)
-	}
-	t := newToken()
-	go func() {
-		for _, token := range tokens {
-			token.Wait()
-			if token.Error() != nil {
-				fmt.Println(token.Error())
-				t.err = token.Error()
-			}
-		}
-		t.flowComplete()
-	}()
-	return t
-}
-
-func flatten(prefix, sep string, in, out map[string]interface{}) {
-	for k, v := range in {
-		key := prefix + sep + k
-		if prefix == "" {
-			key = k
-		}
-		out[key] = v
-		if next, ok := v.(map[string]interface{}); ok {
-			flatten(key, sep, next, out)
-		}
-	}
-}
-
-// SubscribeDeviceUplink subscribes to all uplink messages for the given application and device
-func (c *DefaultClient) SubscribeDeviceUplink(appID string, devID string, handler UplinkHandler) Token {
-	topic := DeviceTopic{appID, devID, Uplink, ""}
-	return c.subscribe(topic.String(), func(mqtt MQTT.Client, msg MQTT.Message) {
-		// Determine the actual topic
-		topic, err := ParseDeviceTopic(msg.Topic())
-		if err != nil {
-			c.ctx.Warnf("Received message on invalid uplink topic: %s", msg.Topic())
-			return
-		}
-
-		// Unmarshal the payload
-		dataUp := &UplinkMessage{}
-		err = json.Unmarshal(msg.Payload(), dataUp)
-		dataUp.AppID = topic.AppID
-		dataUp.DevID = topic.DevID
-
-		if err != nil {
-			c.ctx.Warnf("Could not unmarshal uplink (%s).", err.Error())
-			return
-		}
-
-		// Call the uplink handler
-		handler(c, topic.AppID, topic.DevID, *dataUp)
-	})
-}
-
-// SubscribeAppUplink subscribes to all uplink messages for the given application
-func (c *DefaultClient) SubscribeAppUplink(appID string, handler UplinkHandler) Token {
-	return c.SubscribeDeviceUplink(appID, "", handler)
-}
-
-// SubscribeUplink subscribes to all uplink messages that the current user has access to
-func (c *DefaultClient) SubscribeUplink(handler UplinkHandler) Token {
-	return c.SubscribeDeviceUplink("", "", handler)
-}
-
-// UnsubscribeDeviceUplink unsubscribes from the uplink messages for the given application and device
-func (c *DefaultClient) UnsubscribeDeviceUplink(appID string, devID string) Token {
-	topic := DeviceTopic{appID, devID, Uplink, ""}
-	return c.unsubscribe(topic.String())
-}
-
-// UnsubscribeAppUplink unsubscribes from the uplink messages for the given application
-func (c *DefaultClient) UnsubscribeAppUplink(appID string) Token {
-	return c.UnsubscribeDeviceUplink(appID, "")
-}
-
-// UnsubscribeUplink unsubscribes from the uplink messages that the current user has access to
-func (c *DefaultClient) UnsubscribeUplink() Token {
-	return c.UnsubscribeDeviceUplink("", "")
-}
-
-// PublishDownlink publishes a downlink message
-func (c *DefaultClient) PublishDownlink(dataDown DownlinkMessage) Token {
-	topic := DeviceTopic{dataDown.AppID, dataDown.DevID, Downlink, ""}
-	dataDown.AppID = ""
-	dataDown.DevID = ""
-	msg, err := json.Marshal(dataDown)
-	if err != nil {
-		return &simpleToken{fmt.Errorf("Unable to marshal the message payload")}
-	}
-	return c.mqtt.Publish(topic.String(), QoS, false, msg)
-}
-
-// SubscribeDeviceDownlink subscribes to all downlink messages for the given application and device
-func (c *DefaultClient) SubscribeDeviceDownlink(appID string, devID string, handler DownlinkHandler) Token {
-	topic := DeviceTopic{appID, devID, Downlink, ""}
-	return c.subscribe(topic.String(), func(mqtt MQTT.Client, msg MQTT.Message) {
-		// Determine the actual topic
-		topic, err := ParseDeviceTopic(msg.Topic())
-		if err != nil {
-			c.ctx.Warnf("Received message on invalid downlink topic: %s", msg.Topic())
-			return
-		}
-
-		// Unmarshal the payload
-		dataDown := &DownlinkMessage{}
-		err = json.Unmarshal(msg.Payload(), dataDown)
-		if err != nil {
-			c.ctx.Warnf("Could not unmarshal downlink (%s).", err.Error())
-			return
-		}
-		dataDown.AppID = topic.AppID
-		dataDown.DevID = topic.DevID
-
-		// Call the Downlink handler
-		handler(c, topic.AppID, topic.DevID, *dataDown)
-	})
-}
-
-// SubscribeAppDownlink subscribes to all downlink messages for the given application
-func (c *DefaultClient) SubscribeAppDownlink(appID string, handler DownlinkHandler) Token {
-	return c.SubscribeDeviceDownlink(appID, "", handler)
-}
-
-// SubscribeDownlink subscribes to all downlink messages that the current user has access to
-func (c *DefaultClient) SubscribeDownlink(handler DownlinkHandler) Token {
-	return c.SubscribeDeviceDownlink("", "", handler)
-}
-
-// UnsubscribeDeviceDownlink unsubscribes from the downlink messages for the given application and device
-func (c *DefaultClient) UnsubscribeDeviceDownlink(appID string, devID string) Token {
-	topic := DeviceTopic{appID, devID, Downlink, ""}
-	return c.unsubscribe(topic.String())
-}
-
-// UnsubscribeAppDownlink unsubscribes from the downlink messages for the given application
-func (c *DefaultClient) UnsubscribeAppDownlink(appID string) Token {
-	return c.UnsubscribeDeviceDownlink(appID, "")
-}
-
-// UnsubscribeDownlink unsubscribes from the downlink messages that the current user has access to
-func (c *DefaultClient) UnsubscribeDownlink() Token {
-	return c.UnsubscribeDeviceDownlink("", "")
-}
-
-// PublishActivation publishes an activation
-func (c *DefaultClient) PublishActivation(activation Activation) Token {
-	topic := DeviceTopic{activation.AppID, activation.DevID, Activations, ""}
-	activation.AppID = ""
-	activation.DevID = ""
-	msg, err := json.Marshal(activation)
-	if err != nil {
-		return &simpleToken{fmt.Errorf("Unable to marshal the message payload")}
-	}
-	return c.mqtt.Publish(topic.String(), QoS, false, msg)
-}
-
-// SubscribeDeviceActivations subscribes to all activations for the given application and device
-func (c *DefaultClient) SubscribeDeviceActivations(appID string, devID string, handler ActivationHandler) Token {
-	topic := DeviceTopic{appID, devID, Activations, ""}
-	return c.subscribe(topic.String(), func(mqtt MQTT.Client, msg MQTT.Message) {
-		// Determine the actual topic
-		topic, err := ParseDeviceTopic(msg.Topic())
-		if err != nil {
-			c.ctx.Warnf("Received message on invalid activations topic: %s", msg.Topic())
-			return
-		}
-
-		// Unmarshal the payload
-		activation := &Activation{}
-		err = json.Unmarshal(msg.Payload(), activation)
-		if err != nil {
-			c.ctx.Warnf("Could not unmarshal activation (%s).", err.Error())
-			return
-		}
-		activation.AppID = topic.AppID
-		activation.DevID = topic.DevID
-
-		// Call the Activation handler
-		handler(c, topic.AppID, topic.DevID, *activation)
-	})
-}
-
-// SubscribeAppActivations subscribes to all activations for the given application
-func (c *DefaultClient) SubscribeAppActivations(appID string, handler ActivationHandler) Token {
-	return c.SubscribeDeviceActivations(appID, "", handler)
-}
-
-// SubscribeActivations subscribes to all activations that the current user has access to
-func (c *DefaultClient) SubscribeActivations(handler ActivationHandler) Token {
-	return c.SubscribeDeviceActivations("", "", handler)
-}
-
-// UnsubscribeDeviceActivations unsubscribes from the activations for the given application and device
-func (c *DefaultClient) UnsubscribeDeviceActivations(appID string, devID string) Token {
-	topic := DeviceTopic{appID, devID, Activations, ""}
-	return c.unsubscribe(topic.String())
-}
-
-// UnsubscribeAppActivations unsubscribes from the activations for the given application
-func (c *DefaultClient) UnsubscribeAppActivations(appID string) Token {
-	return c.UnsubscribeDeviceActivations(appID, "")
-}
-
-// UnsubscribeActivations unsubscribes from the activations that the current user has access to
-func (c *DefaultClient) UnsubscribeActivations() Token {
-	return c.UnsubscribeDeviceActivations("", "")
 }
